requirements.txt
django==5.0.8
django-csp==4.0
django-celery-beat==2.8.1
django-phonenumber-field[phonenumbers]
psycopg2-binary==2.9.10
gunicorn==23.0.0
redis==5.2.1
django-redis==6.0.0
ratelimit==2.2.1
python-dotenv==1.0.1
pre-commit==4.5.1